+++ frontend/src/components/pages/HomePage.tsx
@@ -6,14 +6,13 @@ import {
   Heading,
   Icon,
   Spacer,
-  Text,
-  Center,
   VStack,
 } from "@chakra-ui/react";
 import { FilePlus, Search, ArrowRight } from "react-feather";
 import CustomInput from "../common/CustomInput";
 import IntakeHeader from "../intake/IntakeHeader";
 import CaseStatus from "../../types/CaseTypes";
+import FilteredSection from "../dashboard/FilteredSection";
 
 const SecondaryHeader = (): React.ReactElement => {
   return (
@@ -41,41 +40,6 @@ const SecondaryHeader = (): React.ReactElement => {
           New case
         </Button>
       </Flex>
-    </Box>
-  );
-};
-
-const FilteredSection = ({
-  status,
-  cases,
-}: {
-  status: string;
-  cases: [];
-}): React.ReactElement => {
-  return (
-    <Box height="25vh" minHeight="fit-content">
-      <Flex>
-        <Heading textStyle="header-medium">{status}</Heading>
-        <Spacer />
-        <Button
-          variant="tertiary"
-          rightIcon={<Icon as={ArrowRight} />}
-          onClick={() => {}}
-        >
-          View All
-        </Button>
-      </Flex>
-      <Box width="100%" height="100%">
-        {cases.length <= 0 ? (
-          <Center height="full">
-            <Text textStyle="text-medium">
-              No current {status.toLowerCase()} cases
-            </Text>
-          </Center>
-        ) : (
-          <></>
-        )}
-      </Box>
     </Box>
   );
 };
